requirements.txt
flask
pandas
scikit-learn
numpy
requests
beautifulsoup4
pillow
pytesseract
tensorflow==2.15.0
